requirements.txt
python-dotenv
mercadopago
Flask
Flask-Cors